--- src/main/kotlin/com/my/sparta/concert/aggregate/user/application/domain/Payment.kt
@@ -1,3 +1,0 @@
-package com.my.sparta.concert.aggregate.user.application.domain
-
-class Payment
