--- src/main/kotlin/com/my/sparta/concert/aggregate/user/application/domain/Users.kt
@@ -1,8 +1,0 @@
-package com.my.sparta.concert.aggregate.user.application.domain
-
-class Users(
-    var userId: String,
-    var username: String,
-    var age: Int,
-    var wallet: Wallet,
-)
