+++ src/main/kotlin/com/my/sparta/concert/aggregate/user/application/domain/model/Payment.kt
@@ -1,0 +1,3 @@
+package com.my.sparta.concert.aggregate.user.application.domain.model
+
+class Payment
